Refactor HomeScreen with Scaffold and TopAppBar
This commit refactors the `HomeScreen` composable to utilize a `Scaffold` for layout structure.

It introduces a `TopBar` composable, implemented with `CenterAlignedTopAppBar`, to display the title "NZ Events".

The main content, which is the list of events, is now within a `Content` composable that receives padding from the `Scaffold`.

Adjustments to padding have been made in the `Content` composable for better visual alignment.